+++ packages/workspace-fusion-modules/src/lib/modules/fusionBookmarks/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useApplyBookmark';
